# openapi.yaml
openapi: 3.0.3
info:
  title: MicroFoundry Admin API
  description: |
    REST API for the MicroFoundry PaaS platform — a lightweight CloudFoundry-style
    developer experience built on Kubernetes.

    The API is served by the `mf admin` server (default port 8080) alongside the
    HTMX-based admin dashboard. All JSON endpoints live under `/api/`.

    **Authentication**: When OIDC is enabled, API requests require a valid session
    cookie or Bearer token obtained via the `/auth/login` OAuth2 flow. When auth
    is disabled (default for local dev), all endpoints are publicly accessible.
  version: 0.1.0
  contact:
    name: MicroFoundry
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://admin.cf-local.dev:8443
    description: Local TLS (with Keycloak)
...
paths:
...
  /api/apps:
    get:
      tags: [Applications]
      summary: List all applications
      parameters:
        - name: state
          in: query
          schema:
            type: string
            enum: [all, started, stopped]
          description: Filter by application state
      responses:
        "200":
          description: Array of applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AppListItem"
  /api/apps/{name}:
    get:
      tags: [Applications]
      summary: Get application detail
      parameters:
        - $ref: "#/components/parameters/AppName"
      responses:
        "200":
          description: Full application detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppDetail"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Applications]
      summary: Delete an application
      parameters:
        - $ref: "#/components/parameters/AppName"
      responses:
        "204":
          description: Application deleted
        "404":
          $ref: "#/components/responses/NotFound"
  /api/apps/{name}/scale:
    post:
      tags: [Applications]
      summary: Scale application instances, memory, or disk
      parameters:
        - $ref: "#/components/parameters/AppName"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                instances:
                  type: integer
                  minimum: 0
                  maximum: 20
                memory_mb:
                  type: integer
                  minimum: 64
                  maximum: 8192
                disk_mb:
                  type: integer
                  minimum: 256
                  maximum: 8192
      responses:
        "200":
          description: Scale result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScaleResult"
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /api/apps/{name}/observability:
    get:
      tags: [Applications]
      summary: Combined RED metrics and recent logs
      parameters:
        - $ref: "#/components/parameters/AppName"
        - name: duration
          in: query
          schema:
            type: string
            default: "15m"
          description: Log duration window (e.g. 15m, 1h)
      responses:
        "200":
          description: Observability data
          content:
            application/json:
              schema:
                type: object
  /api/apps/{name}/logs/history:
    get:
      tags: [Applications]
      summary: Get historical log lines from Loki
      parameters:
        - $ref: "#/components/parameters/AppName"
        - name: duration
          in: query
          schema:
            type: string
            default: "1h"
      responses:
        "200":
          description: Log history
          content:
            application/json:
              schema:
                type: object

  # ─── Services ──────────────────────────────────────────────────────────────
...
  /api/services/{name}:
    get:
      tags: [Services]
      summary: Get service instance detail (credentials redacted)
      parameters:
        - $ref: "#/components/parameters/ServiceName"
      responses:
        "200":
          description: Service instance detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceInstance"
        "404":
          $ref: "#/components/responses/NotFound"

  # ─── Catalog ───────────────────────────────────────────────────────────────
...
  /api/secrets:
    get:
      tags: [Secrets]
      summary: List all secrets
      responses:
        "200":
          description: Array of secrets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SecretListItem"
    post:
      tags: [Secrets]
      summary: Create a user-defined secret
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, data]
              properties:
                name:
                  type: string
                  pattern: "^[a-z][a-z0-9-]{0,40}[a-z0-9]$"
                description:
                  type: string
                data:
                  type: object
                  additionalProperties:
                    type: string
      responses:
        "201":
          description: Secret created
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/secrets/{name}:
    get:
      tags: [Secrets]
      summary: Get secret detail (values redacted)
      parameters:
        - $ref: "#/components/parameters/SecretName"
      responses:
        "200":
          description: Secret detail with keys listed but values masked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SecretDetail"
        "404":
          $ref: "#/components/responses/NotFound"

  # ─── Clusters ──────────────────────────────────────────────────────────────
...
  /api/settings/registry:
    put:
      tags: [Settings]
      summary: Save container registry settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistryConfig"
      responses:
        "200":
          $ref: "#/components/responses/StatusSaved"
  /api/settings/webhooks:
    get:
      tags: [Settings]
      summary: List configured webhooks
      responses:
        "200":
          description: Webhook list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WebhookConfig"
    post:
      tags: [Settings]
      summary: Create a webhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, url]
              properties:
                name:
                  type: string
                url:
                  type: string
                  format: uri
                events:
                  type: array
                  items:
                    type: string
                    enum:
                      - app.deployed
                      - app.crashed
                      - app.scaled
                      - app.deleted
                      - alert.fired
                      - alert.resolved
                      - service.created
                      - service.deleted
                enabled:
                  type: boolean
                  default: true
      responses:
        "201":
          description: Webhook created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookConfig"
  /api/settings/webhooks/{id}:
    delete:
      tags: [Settings]
      summary: Delete a webhook
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Webhook deleted
  /api/settings/smtp:
    put:
      tags: [Settings]
      summary: Save SMTP settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SMTPConfig"
      responses:
        "200":
          $ref: "#/components/responses/StatusSaved"
  /api/settings/endpoints:
    get:
      tags: [Settings]
      summary: Get discovered service endpoints
      responses:
        "200":
          description: Endpoint list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ServiceEndpoint"
    put:
      tags: [Settings]
      summary: Save endpoint URL overrides
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                overrides:
                  type: object
                  additionalProperties:
                    type: string
      responses:
        "200":
          $ref: "#/components/responses/StatusSaved"

  # ─── Monitoring ────────────────────────────────────────────────────────────
  /api/monitoring/alerts:
    get:
      tags: [Monitoring]
      summary: List active Alertmanager alerts
      responses:
        "200":
          description: Active alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
        "503":
          description: Alertmanager unavailable
...
  /api/orgs:
    get:
      tags: [Organizations]
      summary: List user's organizations
      responses:
        "200":
          description: Organization list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Organization"
    post:
      tags: [Organizations]
      summary: Create an organization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
      responses:
        "201":
          description: Organization created
  /api/orgs/{id}:
    get:
      tags: [Organizations]
      summary: Get organization detail with members
      parameters:
        - $ref: "#/components/parameters/OrgID"
      responses:
        "200":
          description: Organization detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
    delete:
      tags: [Organizations]
      summary: Delete an organization
      parameters:
        - $ref: "#/components/parameters/OrgID"
      responses:
        "204":
          description: Organization deleted
  /api/orgs/{id}/members:
    get:
      tags: [Organizations]
      summary: List organization members
      parameters:
        - $ref: "#/components/parameters/OrgID"
      responses:
        "200":
          description: Member list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Member"
    post:
      tags: [Organizations]
      summary: Add a member to the organization
      parameters:
        - $ref: "#/components/parameters/OrgID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberInvite"
      responses:
        "201":
          description: Member added
  /api/orgs/{id}/members/{email}:
    delete:
      tags: [Organizations]
      summary: Remove a member from the organization
      parameters:
        - $ref: "#/components/parameters/OrgID"
        - name: email
          in: path
          required: true
          schema:
            type: string
            format: email
      responses:
        "204":
          description: Member removed

  # ─── Workspaces ────────────────────────────────────────────────────────────
  /api/workspaces:
    get:
      tags: [Workspaces]
      summary: List user's workspaces
      responses:
        "200":
          description: Workspace list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Workspace"
    post:
      tags: [Workspaces]
      summary: Create a workspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
      responses:
        "201":
          description: Workspace created
  /api/workspaces/{id}:
    get:
      tags: [Workspaces]
      summary: Get workspace detail with members
      parameters:
        - $ref: "#/components/parameters/WorkspaceID"
      responses:
        "200":
          description: Workspace detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
    delete:
      tags: [Workspaces]
      summary: Delete a workspace
      parameters:
        - $ref: "#/components/parameters/WorkspaceID"
      responses:
        "204":
          description: Workspace deleted
  /api/workspaces/{id}/members:
    get:
      tags: [Workspaces]
      summary: List workspace members
      parameters:
        - $ref: "#/components/parameters/WorkspaceID"
      responses:
        "200":
          description: Member list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Member"
    post:
      tags: [Workspaces]
      summary: Add a member to the workspace
      parameters:
        - $ref: "#/components/parameters/WorkspaceID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberInvite"
      responses:
        "201":
          description: Member added
  /api/workspaces/{id}/members/{email}:
    delete:
      tags: [Workspaces]
      summary: Remove a member from the workspace
      parameters:
        - $ref: "#/components/parameters/WorkspaceID"
        - name: email
          in: path
          required: true
          schema:
            type: string
            format: email
      responses:
        "204":
          description: Member removed
  /api/workspaces/{id}/orgs:
    get:
      tags: [Workspaces]
      summary: List organizations in workspace
      parameters:
        - $ref: "#/components/parameters/WorkspaceID"
      responses:
        "200":
          description: Organization list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Organization"

  # ─── Users (Keycloak) ─────────────────────────────────────────────────────
...
components:
...
  parameters:
    AppName:
      name: name
      in: path
      required: true
      schema:
        type: string
      description: Application name
...
  schemas:
...
    AppListItem:
      type: object
      properties:
        name:
          type: string
        organization:
          type: string
        owner:
          type: string
        state:
          type: string
          enum: [STARTED, STOPPED, UPLOADING, STAGING, DEPLOYING, FAILED]
        lifecycle_type:
          type: string
          enum: [buildpack, docker, cnb]
        running_count:
          type: integer
        total_count:
          type: integer
        memory_mb:
          type: integer
        routes:
          type: array
          items:
            type: string
        created_at:
          type: string
...